issue #126 refactor: Separate login form component
