+++ src/pages/LoginPage/LoginPage.jsx
@@ -1,29 +1,22 @@
-import React, { useEffect, useState } from "react";
+import React, { useEffect } from "react";
 import { useDispatch, useSelector } from "react-redux";
 import { useNavigate } from "react-router-dom";
 import { toast } from "react-toastify";
 
 import Logo from "@/assets/img/img-selody-logo/3x.png";
+import LoginForm from "@/components/Login/LoginForm/LoginForm";
 import Google from "@/components/sign/Google";
 import Naver from "@/components/sign/Naver";
-import { login, naverLogin } from "@/features/auth/auth-service.js";
+import { naverLogin } from "@/features/user/user-service.js";
 import useNaver from "@/hooks/useNaver.jsx";
 
 import {
 	ContainerDiv,
 	LeftSideDiv,
 	RightSideDiv,
-	LoginForm,
-	Input,
-	LoginButton,
-	SignUpButton,
 	DividerHr,
 	SocialLoginContainerDiv,
-	EmailSaveLabel,
-	FindPasswordDiv,
 	SocialLoginBtnContainerDiv,
-	LoginAssistanceDiv,
-	InputContainerDiv,
 } from "./LoginPage.styles";
 
 function LoginPage() {
@@ -33,45 +26,6 @@ function LoginPage() {
 	const naverInfo = useNaver();
 
 	const { user } = useSelector((state) => state.auth);
-
-	const [formValue, setFormValue] = useState({ email: "", password: "" });
-
-	const { email, password } = formValue;
-
-	const handleFormValue = (e) => {
-		const changed = {
-			...formValue,
-			[e.target.name]: e.target.value,
-		};
-		setFormValue(changed);
-	};
-
-	const validate = () => {
-		if (!email.trim() || !password.trim()) {
-			toast.error("이메일과 비밀번호는 반드시 입력되어야 합니다.");
-			return false;
-		}
-
-		const emailRegex = /^[\w-.]+@([\w-]+\.)+[\w-]{2,4}$/;
-		if (!emailRegex.test(email)) {
-			toast.error("올바른 이메일 형식이 아닙니다. 다시 입력해주세요.");
-			return false;
-		}
-
-		return true;
-	};
-
-	const handleSubmit = async (event) => {
-		event.preventDefault();
-
-		if (!validate()) return;
-
-		const response = await dispatchFn(login({ email, password }));
-
-		if (response.error) {
-			toast.error("이메일 또는 비밀번호가 잘못되었습니다.");
-		}
-	};
 
 	useEffect(() => {
 		if (user) {
@@ -97,39 +51,7 @@ function LoginPage() {
 				</h3>
 			</LeftSideDiv>
 			<RightSideDiv>
-				<LoginForm onSubmit={handleSubmit}>
-					<h1>LOGIN.</h1>
-					<InputContainerDiv>
-						<Input
-							data-testid="email-input"
-							type="text"
-							name="email"
-							placeholder="이메일을 입력해주세요."
-							onChange={handleFormValue}
-						/>
-						<Input
-							data-testid="password-input"
-							type="password"
-							name="password"
-							placeholder="비밀번호를 입력해주세요."
-							onChange={handleFormValue}
-						/>
-					</InputContainerDiv>
-					<LoginAssistanceDiv>
-						<EmailSaveLabel>
-							<input type="checkbox" id="hidden-checkbox" />
-							<div id="shown-checkbox" />
-							<span>이메일 저장</span>
-						</EmailSaveLabel>
-						<FindPasswordDiv>
-							<span>비밀번호를 잊으셨나요?</span>
-						</FindPasswordDiv>
-					</LoginAssistanceDiv>
-					<LoginButton type="submit">로그인</LoginButton>
-					<SignUpButton type="button" onClick={() => navigate("/signup")}>
-						회원가입
-					</SignUpButton>
-				</LoginForm>
+				<LoginForm />
 				<DividerHr />
 				<SocialLoginContainerDiv>
 					<p>간편 로그인</p>
